fixed env_vars not raising key errors on missing keys
