+++ community/langchain_community/utilities/tableau/simple_datasource_qa.py
@@ -1,7 +1,9 @@
+import os
 import json
 import re
-from typing import Dict, Optional
 import logging
+from typing import Dict, Optional
+from dotenv import load_dotenv
 
 from community.langchain_community.utilities.tableau.vizql_data_service import query_vds, query_vds_metadata
 from community.langchain_community.utilities.tableau.utils import json_to_markdown_table
@@ -111,3 +113,49 @@ def prepare_prompt_inputs(data: dict, user_string: str) -> dict:
         "data_table": data.get('data_table', ''),
         "user_input": user_string
     }
+
+
+def env_vars_simple_datasource_qa(
+    domain=None,
+    site=None,
+    jwt_client_id=None,
+    jwt_secret_id=None,
+    jwt_secret=None,
+    tableau_api_version=None,
+    tableau_user=None,
+    datasource_luid=None,
+    tooling_llm_model=None
+):
+    """
+    Retrieves Tableau configuration from environment variables if not provided as arguments.
+
+    Args:
+        domain (str, optional): Tableau domain
+        site (str, optional): Tableau site
+        jwt_client_id (str, optional): JWT client ID
+        jwt_secret_id (str, optional): JWT secret ID
+        jwt_secret (str, optional): JWT secret
+        tableau_api_version (str, optional): Tableau API version
+        tableau_user (str, optional): Tableau user
+        datasource_luid (str, optional): Datasource LUID
+        tooling_llm_model (str, optional): Tooling LLM model
+
+    Returns:
+        dict: A dictionary containing all the configuration values
+    """
+    # Load environment variables before accessing them
+    load_dotenv()
+
+    config = {
+        'domain': domain if isinstance(domain, str) and domain else os.environ['TABLEAU_DOMAIN'],
+        'site': site or os.environ['TABLEAU_SITE'],
+        'jwt_client_id': jwt_client_id or os.environ['TABLEAU_JWT_CLIENT_ID'],
+        'jwt_secret_id': jwt_secret_id or os.environ['TABLEAU_JWT_SECRET_ID'],
+        'jwt_secret': jwt_secret or os.environ['TABLEAU_JWT_SECRET'],
+        'tableau_api_version': tableau_api_version or os.environ['TABLEAU_API_VERSION'],
+        'tableau_user': tableau_user or os.environ['TABLEAU_USER'],
+        'datasource_luid': datasource_luid or os.environ['DATASOURCE_LUID'],
+        'tooling_llm_model': tooling_llm_model or os.environ['TOOLING_MODEL']
+    }
+
+    return config
